Create Traffic Analysis in Minneapolis
The goal of this project was to build a few models to
represent traffic volume on this busy highway, particularly based on the
hours of the day and day of the week. I wanted to get healthy models
built on just this particular set of variables because I was hoping to
isolate the affect of Rain and Cloudiness on weather.